# application/src/jreleaser/templates/article/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Ublhub API",
    "description" : "Ublhub REST API definitions for you to consume.",
    "termsOfService" : "https://github.com/project-openubl/ublhub/blob/master/CODE_OF_CONDUCT.md",
    "contact" : {
      "name" : "Ublhub API Support",
      "url" : "https://project-openubl.github.io/",
      "email" : "projectopenubl@googlegroups.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "2.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
  "paths" : {
...
    "/api/api/projects" : {
      "get" : {
        "tags" : [ "Project Resource" ],
        "summary" : "List projects",
        "description" : "List all projects",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        } ]
      },
      "post" : {
        "tags" : [ "Project Resource" ],
        "summary" : "Create project",
        "description" : "Create a project",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        } ]
      }
    },
...
    "/api/api/projects/{projectId}/companies/{companyId}/components" : {
      "get" : {
        "tags" : [ "Keys Resource" ],
        "summary" : "Get company components",
        "description" : "List of components",
        "parameters" : [ {
          "name" : "companyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "name",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "parent",
          "in" : "query",
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "type",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ComponentDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:read" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:read" ]
        } ]
      },
      "post" : {
        "tags" : [ "Keys Resource" ],
        "summary" : "Create a company component",
        "description" : "Create component",
        "parameters" : [ {
          "name" : "companyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ComponentDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ComponentDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        } ]
      }
    },
    "/api/api/projects/{projectId}/companies/{companyId}/components/{componentId}" : {
      "get" : {
        "tags" : [ "Keys Resource" ],
        "summary" : "Get company component",
        "description" : "Get one component",
        "parameters" : [ {
          "name" : "companyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "componentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ComponentDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:read" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:read" ]
        } ]
      },
      "put" : {
        "tags" : [ "Keys Resource" ],
        "summary" : "Update company component",
        "description" : "Update a component",
        "parameters" : [ {
          "name" : "companyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "componentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ComponentDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ComponentDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        } ]
      },
      "delete" : {
        "tags" : [ "Keys Resource" ],
        "summary" : "Delete a company component",
        "description" : "Delete a component",
        "parameters" : [ {
          "name" : "companyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "componentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        } ]
      }
    },
    "/api/api/projects/{projectId}/companies/{companyId}/keys" : {
      "get" : {
        "tags" : [ "Keys Resource" ],
        "summary" : "Get company keys",
        "description" : "List of keys",
        "parameters" : [ {
          "name" : "companyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/KeysMetadataRepresentation"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:read" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:read" ]
        } ]
      }
    },
...
    "/api/api/projects/{projectId}/keys" : {
      "get" : {
        "tags" : [ "Keys Resource" ],
        "summary" : "Get project keys",
        "description" : "List of keys",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/KeysMetadataRepresentation"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:read" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:read" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ComponentDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "providerId" : {
            "type" : "string"
          },
          "providerType" : {
            "type" : "string"
          },
          "parentId" : {
            "type" : "string"
          },
          "subType" : {
            "type" : "string"
          },
          "config" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        }
      },
...
      "KeysMetadataRepresentation" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "keys" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/KeyMetadataRepresentation"
            }
          }
        }
      },
...
      "ProjectDto" : {
        "required" : [ "name", "sunat" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "maxLength" : 250,
            "type" : "string"
          },
          "sunat" : {
            "$ref" : "#/components/schemas/SunatDto"
          }
        }
      },
...
      "SunatDto" : {
        "required" : [ "facturaUrl", "guiaUrl", "retencionUrl", "username", "password" ],
        "type" : "object",
        "properties" : {
          "facturaUrl" : {
            "type" : "string"
          },
          "guiaUrl" : {
            "type" : "string"
          },
          "retencionUrl" : {
            "type" : "string"
          },
          "username" : {
            "maxLength" : 250,
            "minLength" : 3,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 250,
            "minLength" : 3,
            "type" : "string"
          }
        }
      },
...
    }
  }
}